Weighted Prioritised Truncated Sum

+++ Assets/_Scripts/SOSteeringData.cs
@@ -6,5 +6,5 @@ using UnityEngine;
 public class SOSteeringData : ScriptableObject
 {
 	public float Weight;
-	public float Priority;
+	public int Priority;
 }
